Added tests for authentication + added the Authentication headers to all tests which use authenticated end points

+++ Backend/src/main/java/com/SYSC4806_Group13/SYSC4806_Project/Security/UserPrincipal.java
@@ -1,6 +1,6 @@
 package com.SYSC4806_Group13.SYSC4806_Project.Security;
 
-import com.SYSC4806_Group13.SYSC4806_Project.Model.User;
+import com.SYSC4806_Group13.SYSC4806_Project.Model.DataModel.User;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -12,13 +12,13 @@ import java.util.List;
 import java.util.Map;
 
 public class UserPrincipal implements OAuth2User, UserDetails {
-    private Integer id;
+    private Long id;
     private String email;
     private String password;
     private Collection<? extends GrantedAuthority> authorities;
     private Map<String, Object> attributes;
 
-    public UserPrincipal(Integer id, String email, String password, Collection<? extends GrantedAuthority> authorities) {
+    public UserPrincipal(Long id, String email, String password, Collection<? extends GrantedAuthority> authorities) {
         this.id = id;
         this.email = email;
         this.password = password;
@@ -43,7 +43,7 @@ public class UserPrincipal implements OAuth2User, UserDetails {
         return userPrincipal;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
