+++ Backend/src/main/java/com/SYSC4806_Group13/SYSC4806_Project/Service/UserService.java
@@ -7,5 +7,5 @@ import org.springframework.security.core.userdetails.UsernameNotFoundException;
 public interface UserService {
     UserDetails loadUserByUsername(String email) throws UsernameNotFoundException;
 
-    UserDetails loadUserById(Integer id);
+    UserDetails loadUserById(Long id);
 }
